+++ src/packages/api/credit-memo/services/get-credit-memos-for-date-range.ts
@@ -1,5 +1,5 @@
 // Import the Query Builder
-import { SearchOptions, SearchResponse, type CreditMemo } from '../../../../types/types';
+import type { CreditMemo, CreditMemoOptions, SearchResponse } from '../../../../types/types';
 import { CreditMemoAPI } from '../credit-memo-api';
 
 /**
@@ -13,7 +13,7 @@ export async function getCreditMemosForDateRange(
 	this: CreditMemoAPI,
 	startDate: Date,
 	endDate: Date,
-	options: SearchOptions<CreditMemo> = {},
+	options: CreditMemoOptions = {},
 ): Promise<SearchResponse<CreditMemo>> {
 	// Ensure the Start Date is Before the End Date
 	if (startDate > endDate) throw new Error('Start date must be before end date');
@@ -25,8 +25,8 @@ export async function getCreditMemosForDateRange(
 	queryBuilder.whereLastUpdatedAfter(startDate);
 	queryBuilder.whereLastUpdatedBefore(endDate);
 
-	// Setup the Search Options
-	queryBuilder.setSearchOptions(options);
+	// Setup the Search Options (if provided)
+	if (options.searchOptions) queryBuilder.setSearchOptions(options.searchOptions);
 
 	// Setup the URL
 	const url = queryBuilder.build();