+++ src/packages/api/credit-memo/services/get-updated-credit-memos.ts
@@ -1,5 +1,5 @@
 // Import the Query Builder
-import { SearchOptions, SearchResponse, type CreditMemo } from '../../../../types/types';
+import type { CreditMemo, CreditMemoOptions, SearchResponse } from '../../../../types/types';
 import { CreditMemoAPI } from '../credit-memo-api';
 
 /**
@@ -11,7 +11,7 @@ import { CreditMemoAPI } from '../credit-memo-api';
 export async function getUpdatedCreditMemos(
 	this: CreditMemoAPI,
 	lastUpdatedDate: Date,
-	options: SearchOptions<CreditMemo> = {},
+	options: CreditMemoOptions = {},
 ): Promise<SearchResponse<CreditMemo>> {
 	// Get the Query Builder
 	const queryBuilder = await this.getQueryBuilder();
@@ -19,8 +19,8 @@ export async function getUpdatedCreditMemos(
 	// Setup the Last Updated Date Filter
 	queryBuilder.whereLastUpdatedAfter(lastUpdatedDate);
 
-	// Setup the Search Options
-	queryBuilder.setSearchOptions(options);
+	// Setup the Search Options (if provided)
+	if (options.searchOptions) queryBuilder.setSearchOptions(options.searchOptions);
 
 	// Setup the URL
 	const url = queryBuilder.build();