+++ src/packages/api/estimates/services/get-updated-estimates.ts
@@ -1,5 +1,5 @@
 // Import the Query Builder
-import { SearchOptions, SearchResponse, type Estimate } from '../../../../types/types';
+import type { Estimate, EstimateOptions, SearchResponse } from '../../../../types/types';
 import { EstimateAPI } from '../estimate-api';
 
 /**
@@ -11,7 +11,7 @@ import { EstimateAPI } from '../estimate-api';
 export async function getUpdatedEstimates(
 	this: EstimateAPI,
 	lastUpdatedDate: Date,
-	options: SearchOptions<Estimate> = {},
+	options: EstimateOptions = {},
 ): Promise<SearchResponse<Estimate>> {
 	// Get the Query Builder
 	const queryBuilder = await this.getQueryBuilder();
@@ -19,8 +19,8 @@ export async function getUpdatedEstimates(
 	// Setup the Last Updated Date Filter
 	queryBuilder.whereLastUpdatedAfter(lastUpdatedDate);
 
-	// Setup the Search Options
-	queryBuilder.setSearchOptions(options);
+	// Setup the Search Options (if provided)
+	if (options.searchOptions) queryBuilder.setSearchOptions(options.searchOptions);
 
 	// Setup the URL
 	const url = queryBuilder.build();