+++ src/packages/api/payment/services/get-payment-by-id.ts
@@ -1,5 +1,5 @@
 // Import the Query Builder
-import { SearchOptions, type Payment } from '../../../../types/types';
+import type { Payment, PaymentOptions } from '../../../../types/types';
 import { PaymentAPI } from '../payment-api';
 
 /**
@@ -8,15 +8,15 @@ import { PaymentAPI } from '../payment-api';
  * @param id - The ID of the payment
  * @returns The Payment
  */
-export async function getPaymentById(this: PaymentAPI, id: string, options: SearchOptions<Payment> = {}): Promise<Payment> {
+export async function getPaymentById(this: PaymentAPI, id: string, options: PaymentOptions = {}): Promise<Payment> {
 	// Get the Query Builder
 	const queryBuilder = await this.getQueryBuilder();
 
 	// Setup ID Filter
 	queryBuilder.whereId(id);
 
-	// Setup the Search Options
-	queryBuilder.setSearchOptions(options);
+	// Setup the Search Options (if provided)
+	if (options.searchOptions) queryBuilder.setSearchOptions(options.searchOptions);
 
 	// Setup the URL
 	const url = queryBuilder.build();