+++ src/types/types.ts
@@ -5,6 +5,7 @@ export { Endpoints } from './enums/endpoints';
 export { Environment } from './enums/environment';
 export { EstimateStatus } from './enums/estimate-status';
 export { GrantType } from './enums/grant-type';
+export { InvoiceStatus } from './enums/invoice-status';
 export { Query } from './enums/query';
 export { TokenType } from './enums/token-type';
 
@@ -13,8 +14,16 @@ export type { Estimate } from './interfaces/estimate';
 export type { Invoice } from './interfaces/invoice';
 export type { Payment } from './interfaces/payment';
 export type { Customer } from './interfaces/customer';
+export type {
+	InvoiceOptions,
+	EstimateOptions,
+	CustomerOptions,
+	PaymentOptions,
+	AccountOptions,
+	CreditMemoOptions,
+	PreferenceOptions,
+} from './interfaces/options';
 export type { Account } from './interfaces/account';
-
 export type { CreditMemo } from './interfaces/credit-memo';
 export type { Preferences } from './interfaces/preferences';
 export type {